Builds: move send_build_status to builds/tasks.py
Mostly to avoid a circular import in https://github.com/readthedocs/readthedocs.org/pull/7548#discussion_r540571159

Also, make projects/tasks.py more project related and small :D
